moved tests to new mod, active on feature=verify
